LG/bot_cadastro_funcionarios/requirements.txt
beautifulsoup4==4.12.3
botcity-framework-base==0.4.4
botcity-framework-web==0.9.0
botcity-maestro-sdk==0.5.1
certifi==2024.8.30
charset-normalizer==3.3.2
chromedriver==2.24.1
distro==1.9.0
et-xmlfile==1.1.0
idna==3.10
importlib_metadata==8.5.0
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==3.141.0
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==1.26.20
webdriver-manager==4.0.2
zipp==3.20.2
